LJPS-50 change directory
