+++ ljps-frontend/pages/index.js
@@ -1,73 +1,67 @@
-import Head from "next/head";
-import Image from "next/image";
-import styles from "../styles/Home.module.css";
-import Link from "next/link"
+import React, { useState } from "react";
+import { Player } from "@lottiefiles/react-lottie-player";
+import { useRouter } from "next/router";
 
 export default function Home() {
+  const [role, setRole] = useState("User");
+  var router = useRouter();
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    if (role == "User") {
+      router.push("/learning-journey");
+    } else {
+      router.push("/roles");
+    }
+  };
+
   return (
-    <div className={styles.container}>
-      <Head>
-        <title>Create Next App</title>
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
+    <div className="row">
+      <div className="col-md-6 d-flex justify-content-center">
+        <Player
+          src="https://assets3.lottiefiles.com/packages/lf20_9gddhmw6.json"
+          className="player"
+          loop
+          autoplay
+          style={{ height: "550px", width: "550px" }}
+        />
+      </div>
 
-
-      <main className={styles.main}>
-        <h1 className={styles.title}>
-          Welcome to <a href="https://nextjs.org">Next.js!</a>
-        </h1>
-
-
-        <p className={styles.description}>
-          Get started by editing{" "}
-          <code className={styles.code}>pages/index.js</code>
-        </p>
-
-        <div className={styles.grid}>
-          <a href="https://nextjs.org/docs" className={styles.card}>
-            <h2>Documentation &rarr;</h2>
-            <p>Find in-depth information about Next.js features and API.</p>
-          </a>
-
-          <a href="https://nextjs.org/learn" className={styles.card}>
-            <h2>Learn &rarr;</h2>
-            <p>Learn about Next.js in an interactive course with quizzes!</p>
-          </a>
-
-          <a
-            href="https://github.com/vercel/next.js/tree/canary/examples"
-            className={styles.card}
+      <div className="col-md-6 d-flex justify-content-center flex-column px-5">
+        <div className="d-flex mb-4">
+          <button
+            type="button"
+            className="btn btn-primary btn-circle d-flex justify-content-center align-items-center"
+            style={{ width: "30px", height: "30px", borderRadius: "15px" }}
           >
-            <h2>Examples &rarr;</h2>
-            <p>Discover and deploy boilerplate example Next.js projects.</p>
-          </a>
+            <i className="bi bi-printer"></i>
+          </button>
 
-          <a
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-          >
-            <h2>Deploy &rarr;</h2>
-            <p>
-              Instantly deploy your Next.js site to a public URL with Vercel.
-            </p>
-          </a>
+          <h4 className="ms-3">
+            <span className="text-primary fw-bold"> All-In-One </span> Learning
+            Journey Planning System
+          </h4>
         </div>
-      </main>
 
-      <footer className={styles.footer}>
-        <a
-          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Powered by{" "}
-          <span className={styles.logo}>
-            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-          </span>
-        </a>
-      </footer>
+        <p className="mb-3"> Select your Role to proceed</p>
 
+        <form onSubmit={handleSubmit}>
+          <select
+            className="form-select mb-4"
+            onChange={(e) => setRole(e.target.value)}
+          >
+            <option value="User">User</option>
+            <option value="Manager">Manager</option>
+            <option value="Admin">Admin</option>
+          </select>
 
+          <div className="mt-5 d-flex justify-content-end">
+            <button type="submit" className="btn btn-outline-primary">
+              Submit
+            </button>
+          </div>
+        </form>
+      </div>
     </div>
   );
 }
